requirements.txt
google-genai
streamlit
faiss-cpu
sentence-transformers
numpy
python-dotenv
